types: remove shadow interface of MessageView
Now that we use true private syntax in MessageView, the extra interface for TS is no longer necessary
